Fixed users roles bugs

+++ WebApp2/ClientApp/src/app/services/login.service.ts
@@ -15,20 +15,35 @@ export class LoginService {
   }
 
   logIn(user: User) {
-    if (user.role === 'admin') {
-      this.router.navigate(['/admin']);
-      localStorage.setItem('role-token', 'admin');
+    if (user != null) {
+
+      if (user.name === 'ErrorUSer') {
+        this.router.navigate(['/log-in']);
+        return confirm('Incorrect login!');
+      } else {
+        localStorage.setItem('cr-user', user.name);
+        if (user.role === 'admin') {
+
+          this.router.navigate(['/admin']);
+          localStorage.setItem('role-token', 'admin');
+
+        } else {
+
+          this.router.navigate(['']);
+          localStorage.setItem('role-token', 'user');
+        }
+      }
     } else {
-      this.router.navigate(['']);
-      localStorage.setItem('role-token', 'user');
+      this.router.navigate(['/log-in']);
     }
   }
 
   logOut() {
     localStorage.setItem('role-token', 'user');
+    localStorage.setItem('cr-user', 'unknown');
   }
 
-  isLogged() {
+  isAdmin() {
     if (localStorage.getItem('role-token') === 'admin') {
       return true;
     } else {
